# docs/swagger.yaml
basePath: /
definitions:
...
  common.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  common.Response:
    properties:
      data: {}
      error:
        $ref: "#/definitions/common.ResponseError"
      meta:
        $ref: "#/definitions/common.ResponseMeta"
    type: object
...
  common.ResponseMeta:
    properties:
      extra:
        additionalProperties: true
        type: object
      pagination:
        $ref: "#/definitions/common.PaginationMeta"
      timestamp:
        type: string
      trace_id:
        type: string
    type: object
...
  orders.CreateOrderRequest:
    properties:
      buyer_organization_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      items:
        items:
          $ref: "#/definitions/orders.CreateOrderItemRequest"
        minItems: 1
        type: array
      metadata:
        additionalProperties: true
        type: object
      notes:
        example: Special delivery instructions
        maxLength: 1000
        type: string
      seller_organization_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      shipping_address:
        $ref: "#/definitions/entities_requests_orders.Address"
    required:
      - buyer_organization_id
      - items
      - seller_organization_id
    type: object
...
  orders.OrderResponse:
    properties:
      actual_delivery_date:
        example: "2024-01-19T14:15:00Z"
        type: string
      buyer_organization_id:
        example: 123e4567-e89b-12d3-a456-426614174001
        type: string
      buyer_user_id:
        example: 123e4567-e89b-12d3-a456-426614174003
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      discount_amount:
        example: 0
        type: number
      estimated_delivery_date:
        example: "2024-01-20T10:30:00Z"
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      items:
        items:
          $ref: "#/definitions/orders.OrderItemResponse"
        type: array
      metadata:
        additionalProperties: true
        type: object
      notes:
        example: Special delivery instructions
        type: string
      order_number:
        example: ORD-2024-001
        type: string
      seller_organization_id:
        example: 123e4567-e89b-12d3-a456-426614174002
        type: string
      shipping_address:
        $ref: "#/definitions/orders.AddressResponse"
      shipping_amount:
        example: 50
        type: number
      status:
        example: pending
        type: string
      status_history:
        items:
          $ref: "#/definitions/orders.OrderStatusHistoryResponse"
        type: array
      subtotal_amount:
        example: 255
        type: number
      tax_amount:
        example: 25.5
        type: number
      total_amount:
        example: 330.5
        type: number
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the KisanLink E-commerce service API.
  title: KisanLink E-commerce API
  version: "1.0"
paths:
...
  /api/v1/orders:
    get:
      consumes:
        - application/json
      description: Retrieve a list of orders with filtering and pagination
      parameters:
        - default: 1
          description: Page number
          in: query
          name: page
          type: integer
        - default: 20
          description: Items per page
          in: query
          name: limit
          type: integer
        - description: Filter by buyer ID
          in: query
          name: buyer_id
          type: string
        - description: Filter by seller ID
          in: query
          name: seller_id
          type: string
        - description: Filter by status
          in: query
          name: status
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    items:
                      $ref: "#/definitions/orders.OrderResponse"
                    type: array
                  meta:
                    allOf:
                      - $ref: "#/definitions/common.ResponseMeta"
                      - properties:
                          pagination:
                            $ref: "#/definitions/common.PaginationMeta"
                        type: object
                type: object
      summary: List orders
      tags:
        - orders
    post:
      consumes:
        - application/json
      description: Create a new order with items
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Order information
          in: body
          name: order
          required: true
          schema:
            $ref: "#/definitions/orders.CreateOrderRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/orders.OrderResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Create a new order
      tags:
        - orders
  /api/v1/orders/{id}:
    get:
      consumes:
        - application/json
      description: Retrieve an order by its ID
      parameters:
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/orders.OrderResponse"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Get order by ID
      tags:
        - orders
    put:
      consumes:
        - application/json
      description: Update order details including status
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
        - description: Order update data
          in: body
          name: order
          required: true
          schema:
            $ref: "#/definitions/orders.UpdateOrderRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/orders.OrderResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Update an order
      tags:
        - orders
  /api/v1/orders/{id}/cancel:
    post:
      consumes:
        - application/json
      description: Cancel an order and release inventory
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    type: string
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "404":
          description: Not Found
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Cancel an order
      tags:
        - orders
  /api/v1/orders/{id}/payment:
    post:
      consumes:
        - application/json
      description: Process payment for an existing order
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
        - description: Payment processing request
          in: body
          name: payment
          required: true
          schema:
            $ref: "#/definitions/orders.ProcessPaymentRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/orders.PaymentResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Process payment for an order
      tags:
        - orders
  /api/v1/orders/{id}/payment/status:
    get:
      consumes:
        - application/json
      description: Retrieve the payment status for an order
      parameters:
        - description: Bearer token
          in: header
          name: Authorization
          required: true
          type: string
        - description: Order ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  data:
                    $ref: "#/definitions/orders.PaymentResponse"
                type: object
        "400":
          description: Bad Request
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
        "403":
          description: Forbidden
          schema:
            allOf:
              - $ref: "#/definitions/common.Response"
              - properties:
                  error:
                    $ref: "#/definitions/common.ResponseError"
                type: object
      summary: Get payment status for an order
      tags:
        - orders
...
swagger: "2.0"